Fix memory alignment problem and seg fault.
When in debug mode, the memory is not aligned and a segfault is obtained when using
an simd operation that expects data to be aligned.

The app crashes mercilesly.

The situation was solved thanks to GBMotta, suffering a day long debugging session with the author of this commit.

We can solve this (and we do) by:
 1- allocating more memory and selecting the correct aligned position.
 2- allocating the array as aligned with the proper compiler keyword. (careful gcc and clang agree but msvc does not).
 3- use the damn simd api and imlement an instruction that does not expect the memory to be aligned.

The faster option is to have the memory aligned as soon as possible. (probably)> option 2.